this application hab been compiled with default appl rom size (144k). it works properly with old eloader and old eupdater. this application uses sys folder at svn revision 18201 and embobj/embobj folder at revision 18200. (so with mc-stuff with few rom)
svn path=/trunk/iCub/firmware/; revision=18205

+++ emBODY/arch-armcmx/apps/tests/app-skeleton-ems-V05/app_extra/EOMappMotorController.c
@@ -231,7 +231,7 @@ static void s_eom_appMotorController_taskStartup(EOMtask *tsk, uint32_t t)
 
 static void s_eom_appMotorController_taskRun(EOMtask *tsk, uint32_t evtmsgper)
 {
-    float *pwm;
+    int16_t *pwm;
     eOevent_t evt;
     uint32_t encoders_values[6];
